refactor: wrapper 파일 이름에 오타 수정
indext => index

#22
